working basic implementation

+++ scripts/fullscreenNotificationService.sh
@@ -3,8 +3,9 @@
 PIPE="$HOME/.config/hypr/pipes/fullscreen_pipe"
 
 # Ensure the pipe exists but first remove it so its flushed
-rm "$PIPE"
-[ -p "$PIPE" ] || mkfifo "$PIPE"
+[ -p "$PIPE" ] && rm "$PIPE"
+mkfifo "$PIPE"
+
 
 echo "Daemon started. Waiting for fullscreen events..."
 
@@ -27,20 +28,22 @@ while true; do
     done
 
     if [[ $registered -eq 1 ]]; then
-        tmpArray=()
-        for item in "${workspacesInFs[@]}"; do
-            if [[ "$item" != "$wsid" ]]; then
-                tmpArray+=("$item")
-            fi
-        done
-        workspacesInFs=("${tmpArray[@]}")    
-        registered=0    
+        if [[ "$event" == "fullscreen" ]]; then
+            tmpArray=()
+            for item in "${workspacesInFs[@]}"; do
+                if [[ "$item" != "$wsid" ]]; then
+                    tmpArray+=("$item")
+                fi
+            done
+            workspacesInFs=("${tmpArray[@]}")    
+            registered=0
+        elif [[ "$event" == "newWindow" ]]; then
+            # Send a notification
+            DISPLAY=:0 DBUS_SESSION_BUS_ADDRESS="unix:path=/run/user/$(id -u)/bus" \
+            notify-send "Your friendly daemon service" "You are in full screen mode but the new window has been created :)" -t 3000
+        fi
     else
         workspacesInFs+=("$wsid")
     fi
-    
-    # Send a notification
-    DISPLAY=:0 DBUS_SESSION_BUS_ADDRESS="unix:path=/run/user/$(id -u)/bus" \
-    notify-send "Hyprland" "Fullscreen event: ${workspacesInFs[*]}" -t 3000
     
 done
